Refactor task list page objects
This moves the page objects in to their own files so they can be used my
multiple pages and tests.
