+++ spec/support/autoload/page_objects/task_list_item.rb
@@ -1,0 +1,6 @@
+module PageObjects
+  class TaskListItem < SitePrism::Section
+    element :link, "a"
+    element :status_tag, ".govuk-tag"
+  end
+end
