+++ spec/support/autoload/page_objects/task_list_section.rb
@@ -1,0 +1,6 @@
+module PageObjects
+  class TaskListSection < SitePrism::Section
+    element :heading, "h2"
+    sections :items, TaskListItem, ".app-task-list__item"
+  end
+end
