+++ spec/support/autoload/page_objects/teacher_interface/application.rb
@@ -1,15 +1,5 @@
 module PageObjects
   module TeacherInterface
-    class TaskListItem < SitePrism::Section
-      element :link, "a"
-      element :status, ".govuk-tag"
-    end
-
-    class TaskListSection < SitePrism::Section
-      element :heading, "h2"
-      sections :task_list_items, TaskListItem, ".app-task-list__item"
-    end
-
     class Application < SitePrism::Page
       set_url "/teacher/application"
 
@@ -20,18 +10,7 @@ module PageObjects
       element :check_answers, ".govuk-button:nth-of-type(1)"
       element :save_and_sign_out, ".govuk-button:nth-of-type(2)"
 
-      sections :task_list_sections, TaskListSection, ".app-task-list>li"
-
-      def find_item(link_text)
-        task_list_sections
-          .flat_map(&:task_list_items)
-          .map(&:link)
-          .find { |link| link.text == link_text }
-      end
-
-      def click_item(link_text)
-        find_item(link_text).click
-      end
+      section :task_list, TaskList, ".app-task-list"
     end
   end
 end
