+++ spec/support/autoload/page_objects/teacher_interface/further_information_requested.rb
@@ -4,6 +4,7 @@ module PageObjects
       set_url "/teacher/application/further_information_requests/{further_information_request_id}"
 
       element :heading, ".govuk-heading-l"
+      section :task_list, TaskList, ".app-task-list"
     end
   end
 end
